projet_SEYNABOU_KALDJOB/tp6/requirements.txt
kivy
kivymd
matplotlib
praw
urllib
urllib.request
_collections
xmltodict
datetime
wordcloud
re
numpy 
math
scipy.sparse 